+++ frontend/CampusSafe/app-example/(tabs)/_layout.tsx
