+++ frontend/CampusSafe/app-example/(tabs)/explore.tsx
