+++ frontend/CampusSafe/app-example/(tabs)/index.tsx
